+++ frontend/src/pages/LinkAnswer.tsx
@@ -10,6 +10,7 @@ import ProgressBar from "../components/ProgressBar";
 import { answerListState } from "../../atoms/AnswerStore";
 import TagPart from "../components/Answer/TagPart";
 import { atom } from "recoil";
+import axios from "axios";
 
 export const currentQuestionIndexState = atom<number>({
   key: "currentQuestionIndexState",
@@ -34,13 +35,59 @@ function LinkAnswer() {
     }
   };
 
-  const handleNextButtonClick = () => {
+  const handleNextButtonClick = async () => {
     if (currentQuestionIndex < questions.questions.length - 1) {
       // 다음 질문으로 이동
       setCurrentQuestionIndex((prevIndex) => prevIndex + 1);
       console.log(answerList);
     } else {
-      // 마지막 질문이면 "/check" 페이지로 이동
+      // 마지막 질문의 경우
+      const feedApiUrl = "http://localhost:8000/api/answers/";
+      const gptApiUrl = "http://localhost:8000/api/feedbacks/summary/";
+
+      // POST 요청할 데이터
+      const postData = {
+        form_id: answerList[0].form_id,
+        category: answerList[0].category,
+        tags_work: answerList[0].tags_work,
+        tags_attitude: answerList[0].tags_attitude,
+        answers: answerList[0].answers,
+      };
+
+      // 답변이 있는지 확인
+      if (answerList && answerList[0].answers.length > 0) {
+        try {
+          const response = await axios.post(feedApiUrl, postData);
+          console.log("답변 제출");
+          console.log(response.data);
+        } catch (error) {
+          // 필요한 경우 에러 처리
+          console.error("폼 없음", error);
+        }
+
+        try {
+          const storedUserid = localStorage.getItem("author_id");
+
+          // axios를 사용해 POST 요청 보내기
+          const response = await axios.post(gptApiUrl, {
+            user_id: storedUserid,
+          });
+
+          // API 응답이 성공인 경우 요약 정보를 상태 변수에 저장
+          if (response.data.status === "success") {
+            console.log("요약 생성 요청 전송 성공");
+          } else {
+            // API 응답이 실패인 경우 에러 메시지 출력
+            console.error("데이터 없음:", response.data.status);
+          }
+        } catch (error) {
+          // 비동기 함수 실행 중 발생한 에러 처리
+          console.error("API 요청 실패:", error);
+        }
+      }
+
+      localStorage.setItem("author_id", "0");
+      // "/LinkFinish" 페이지로 이동
       nextNavigate("/LinkFinish");
     }
   };
@@ -76,9 +123,6 @@ function LinkAnswer() {
           <div className="flex-full">
             <p className="font-pre text-[22px] font-bold text-center px-10">
               {currentQuestion?.context}
-            </p>
-            <p className="font-pre text-[14px] text-gray-400 text-center">
-              키워드를 최대 5개까지 선택해주세요.
             </p>
           </div>
           <div className="flex flex-1 flex-col justify-center items-center">
