+++ frontend/src/pages/LinkMain.tsx
@@ -1,19 +1,33 @@
-import clover from '../assets/clover.svg';
-import ChangePage from '../components/ChangePage';
+import { useEffect, useState } from "react";
+import clover from "../assets/clover.svg";
+import ChangePage from "../components/ChangePage";
 
 function LinkMain() {
+  const [user_id, setUser_id] = useState(0);
+
+  useEffect(() => {
+    // Query Parameters를 가져오기
+    const queryParams = new URLSearchParams(window.location.search);
+
+    // user_id 값을 가져오기
+    const userId = queryParams.get("user_id");
+    if (userId) {
+      setUser_id(parseInt(userId));
+    }
+  }, []);
+
   return (
     <div className="flex justify-center items-center min-h-screen">
-      <ChangePage nextpage="/LinkStart" />
+      <ChangePage nextpage={`/LinkStart?user_id=${1}`} />
       <div
         className="flex h-full flex-col justify-center items-center relative bg-c-emerald bg-opacity-35 min-h-screen w-full sm:w-[393px] lg:w-[393px]"
         // style={{ width: "393px", height: "852px" }}
       >
         <div className="w-full h-full flex flex-1 flex-col justify-center items-center">
-          <p className="text-black dark:text-white text-center font-pre text-[22px] font-bold">
+          <p className="text-black text-center font-pre text-[22px] font-bold">
             안녕하세요!
           </p>
-          <p className="text-black dark:text-white text-center font-pre text-[22px] font-bold">
+          <p className="text-black text-center font-pre text-[22px] font-bold">
             클로피에 오신 걸 환영해요.
           </p>
         </div>
